requirements.txt
Flask
flask-cors
numpy
chromadb
sentence-transformers
python-dotenv
google-genai
opencv-python
whisper
PyPDF2
python-docx
ffmpeg-python
torch
